requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
nuscenes-devkit>=1.1.11
jupyter>=1.0.0